refactor: extract bloated form jsx and its logic
the old `CreateNoteForm` component has gotten very bloated with the form
logic and other display & data logic where a refactor has absolutely
became necessary. this commit extracts the form, formData, tagsText and
all their relevant logic as well as the submit handler for them into a
new component. further refactors will follow.
